# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Hermes Web UI API",
    "description": "BFF server API for Hermes Web UI — chat sessions, scheduled jobs, platform channels, model management, skills, memory, logs, file browser, group chat, and terminal.",
    "version": "0.5.9"
  },
  "servers": [
    {
      "url": "http://localhost:8648",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/api/hermes/sessions/hermes": {
      "get": {
        "tags": [
          "Sessions"
        ],
        "summary": "Get hermes",
        "description": "GET /api/hermes/sessions/hermes",
        "operationId": "listHermesSessions",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/api/hermes/sessions/hermes/{id}": {
      "get": {
        "tags": [
          "Sessions"
        ],
        "summary": "Get :id",
        "description": "GET /api/hermes/sessions/hermes/:id",
        "operationId": "getHermesSession",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/api/hermes/{*any}": {
      "get": {
        "tags": [
          "Proxy"
        ],
        "summary": "Proxy to upstream Hermes API",
        "description": "Forwards unmatched /api/hermes/* requests to upstream Hermes gateway. Supports all upstream endpoints.",
        "operationId": "proxyHermes",
        "responses": {
          "200": {
            "description": "Proxied response from upstream"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "502": {
            "description": "Proxy failure"
          }
        }
      },
      "post": {
        "tags": [
          "Proxy"
        ],
        "summary": "Proxy to upstream Hermes API",
        "description": "Forwards unmatched /api/hermes/* requests to upstream Hermes gateway. Supports all upstream endpoints.",
        "operationId": "proxyHermesPost",
        "responses": {
          "200": {
            "description": "Proxied response from upstream"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "502": {
            "description": "Proxy failure"
          }
        }
      },
      "put": {
        "tags": [
          "Proxy"
        ],
        "summary": "Proxy to upstream Hermes API",
        "description": "Forwards unmatched /api/hermes/* requests to upstream Hermes gateway. Supports all upstream endpoints.",
        "operationId": "proxyHermesPut",
        "responses": {
          "200": {
            "description": "Proxied response from upstream"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "502": {
            "description": "Proxy failure"
          }
        }
      },
      "delete": {
        "tags": [
          "Proxy"
        ],
        "summary": "Proxy to upstream Hermes API",
        "description": "Forwards unmatched /api/hermes/* requests to upstream Hermes gateway. Supports all upstream endpoints.",
        "operationId": "proxyHermesDelete",
        "responses": {
          "200": {
            "description": "Proxied response from upstream"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "502": {
            "description": "Proxy failure"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Token"
      }
    },
...
    "responses": {
      "Unauthorized": {
        "description": "Unauthorized - Invalid or missing authentication token",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Unauthorized"
                }
              }
            }
          }
        }
      },
...
    }
  }
}